requirements.txt
# Core
flask==3.0.0
gunicorn==21.2.0
redis==5.0.1
celery==5.3.6
requests==2.31.0
python-dotenv==1.0.0

# AWS
boto3==1.34.0
botocore==1.34.0

# ML / Data Science
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
tensorflow==2.15.0
joblib==1.3.2
scipy==1.11.4

# Monitoring & Metrics
psutil==5.9.8
prometheus-client==0.19.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
fakeredis==2.20.1
moto==4.2.12

# Utilities
APScheduler==3.10.4
pyyaml==6.0.1
marshmallow==3.20.1